feat: replace DestroyMembership action
